refactor(frontend): split RunningBoard into sub-components
Split RunningBoard (454→1 line) into:
- running-board/ScheduledTodoCard.tsx: scheduled todo card
- running-board/ExecutionRecordCard.tsx: execution record card
- running-board/RunningBoardColumnView.tsx: column view
- running-board/helpers.tsx: formatNextRunAt, COLUMN_ICONS
- running-board/index.tsx: main component

All exports maintained via re-export from original file.
